Markdown formatting
Basic formatting:
 - Whitespaces removal, including newpages (^L) and useless newlines
 - Add newlines where it should be (after section titles etc...)
 - Add section titles hierarchically
 - Format (italic and bold)

The newlines follow now a semantic meaning *for the CAS version*:
this makes much more sense for diff-ing and history checks.
Note that it hasn't been done for the CAT version since it has been automatically
generated and that newlines reflect the original PDF's newlines: for now,
it might make the review process easier.

Some minor content changes for the OCR-generated document:
 - Modify the caught misinterpreted characters
 - Add the missed lines (usually underlined ones)
 - Remove the headers and page numbers